+++ web/src/web/pages/home/home.py
@@ -4,7 +4,7 @@ from dash import html
 
 # noqa suppresses black errors when linting since you need this import for
 # access to callbacks
-import web.pages.sitrep.callbacks  # noqa
+import web.pages.sitrep.callbacks.cytoscape  # noqa
 from web.style import colors
 
 dash.register_page(__name__, path="/", name="Home")
